Fix crash report when SFTP connection is closed very fast
In the SFTP client if the socket is successfully created, but is immediately
closed, even before the code gets to the inet:peername call, there's a huge
crash report. This commit handles this error: adds a new initial state
('uninitialized') to the ssh_connection_handler and the peername call is made
in this state. The start_link sends an event immediately after the FSM start,
which is handled in the new uninitialized state. If the peername call is
successful, the FSM goes to the hello state and everything works as it used to
be. If it's not successful, the FSM terminates. The reason why I introduced
this new state is that I couldn't stop the FSM from the init function without
crash reports, but with a good error value. The rest of the commit is just a
bunch of case statements to propagate the error.